wip: mdl-dialog animations
