+++ src/components/dialog/mdl-dialog-configuration.ts
@@ -1,0 +1,86 @@
+import {
+  ViewContainerRef,
+  Type
+} from '@angular/core';
+
+/**
+ * @deprecated cusotm dialogs no longer need to implement this interface
+ */
+export interface IMdlCustomDialog {
+  viewContainerRef: ViewContainerRef;
+}
+
+/**
+ * The simple Dialog can have as much actions as needed by the user.
+ */
+export interface IMdlDialogAction {
+  /**
+   * the handler is a callback function. this funciton will be called if
+   * the action button was clicked.
+   */
+  handler: () => void;
+  /**
+   * the text of the action button
+   */
+  text: string;
+  /**
+   * is this a closing aciton? means the action is called if the user pressed the esc key.
+   */
+  isClosingAction?: boolean;
+}
+
+/**
+ * Dialog configuration for all dialogs (simple or custom)
+ */
+export interface IMdlDialogConfiguration {
+  /**
+   * The viewcontainerref the dialog will be attached to.
+   * required if not provided by setDefaultViewContainerRef.
+   */
+  vcRef?: ViewContainerRef;
+  /**
+   * true if the dialog should be opened as modal.
+   */
+  isModal?: boolean;
+}
+
+/**
+ * The simple dialog. Easy to use - dosn't need a special component.
+ */
+export interface IMdlSimpleDialogConfiguration extends IMdlDialogConfiguration {
+  /**
+   * the title of the dialog
+   */
+  title?: string;
+  /**
+   * the message that should be displayed
+   */
+  message: string;
+  /**
+   * the actions that are used for this dialog (the order will be reversed by mdl.
+   */
+  actions: [IMdlDialogAction];
+  /**
+   * should the actions be displayed as full width actions. every aciton is one row.
+   */
+  fullWidthAction?: boolean;
+}
+
+/**
+ * Configuration for a custom dialog. You need to provide a component that
+ * should be used as the content of the dialog. the component mus match the
+ * fowllowing conditions:
+ * - must implement IMdlCustomDialog
+ * - must be an entrycompnent (property of your module)
+ * If youn need acces to the MdlDialogReference you may inject it in your constructor:
+ *
+ * export class MyDialog implements IMdlCustomDialog {
+ *
+ *   constructor(private dialogref: MdlDialogReference){}
+ *
+ *   ...
+ * }
+ */
+export interface IMdlCustomDialogConfiguration extends IMdlDialogConfiguration {
+  component: Type<any>;
+}
